// src/app/models/master.ts
export class AuthenticationDetails {
    IsAuth: boolean;
    UserID: Guid;
    UserName: string;
    DisplayName: string;
    EmailAddress: string;
    UserRole: string;
    Token: string;
    ClientID:string;
    Company:string;


    MenuItemNames: string;
    Profile: string;
    RefreahToken: string;
    Expires: string;
    Issued: string;
    Expiresin: string;
    TourStatus: boolean;
}
// ...
export class ActionLog extends CommonClass {
    ID: number;
    UserID: Guid;
    AppName: string;
    Action: string;
    ActionText: string;
    UsedOn: string;
}